Updated professor view and supporting JS.
NOTE: This commit amended to include a more topical message in the "View
Feedback" view in the event that no feedback is returned by the server;
this is caused by a lack of feedback, not "no TA" as was previously
indicated.

Core logic no longer relies on jQuery, and features event-driven
pattern conformity with a moduler pipeline model.

Fixes:
- Large chunks of #12
- #56
- #55
- #54
- Most of #43
- Most of #39
- Most of #38

This commit is expected to conflict with PR#51 because of modifications
to dbhandler.py. PR#51 will be closed as a result, and changes will be
backported onto this branch.